pyproject.toml
[project]
name = "openai-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
]
